+++ src/main/java/com/scaler/productservice/services/ProductService.java
@@ -6,11 +6,11 @@ import com.scaler.productservice.exceptions.NotFoundExceptions;
 import java.util.List;
 
 public interface ProductService {
-    GenericProductDto getProductByID(int id) throws NotFoundExceptions;
+    GenericProductDto getProductByID(long id) throws NotFoundExceptions;
     List<GenericProductDto> getAllProducts();
     GenericProductDto createProduct(GenericProductDto product);
-    GenericProductDto deleteProduct(int id) throws NotFoundExceptions;
-    GenericProductDto updateProduct(int id,GenericProductDto product);
+    GenericProductDto deleteProduct(long id) throws NotFoundExceptions;
+    GenericProductDto updateProduct(long id,GenericProductDto product);
